Finished almost every repository functionality and their unit test except PrescriptionRepository

+++ clinic/clinic/Models/clinic_service.cs
@@ -22,7 +22,7 @@ namespace clinic.Models
     
         public int id { get; set; }
         public string service_name { get; set; }
-        public decimal price { get; set; }
+        public long price { get; set; }
         public Nullable<bool> is_active { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
